refactor(search-app): use searchStateContainerDirective
This implies to create a new component that embed all the search elements. The directive must be applied
outside of the directive otherwise the SearchFacade service is no correctly injected
