+++ apps/search/src/app/app.component.spec.ts
@@ -1,29 +1,9 @@
 import { NO_ERRORS_SCHEMA } from '@angular/core'
 import { async, TestBed } from '@angular/core/testing'
 import { RouterTestingModule } from '@angular/router/testing'
-import { BootstrapService } from '@lib/common'
-import { SearchFacade } from '@lib/search'
 import { EffectsModule } from '@ngrx/effects'
 import { StoreModule } from '@ngrx/store'
-import { of } from 'rxjs'
 import { AppComponent } from './app.component'
-
-const configFacetMock = {
-  mods: {
-    search: {
-      facetConfig: {
-        tag: {},
-      },
-    },
-  },
-}
-const boostrapServiceMock = {
-  uiConfReady: jest.fn(() => of(configFacetMock)),
-}
-const searchFacadeMock = {
-  setConfigAggregations: jest.fn(),
-  requestMoreResults: jest.fn(),
-}
 
 describe('AppComponent', () => {
   beforeEach(async(() => {
@@ -35,16 +15,6 @@ describe('AppComponent', () => {
       ],
       declarations: [AppComponent],
       schemas: [NO_ERRORS_SCHEMA],
-      providers: [
-        {
-          provide: BootstrapService,
-          useValue: boostrapServiceMock,
-        },
-        {
-          provide: SearchFacade,
-          useValue: searchFacadeMock,
-        },
-      ],
     }).compileComponents()
   }))
