+++ apps/search/src/app/app.component.ts
@@ -1,7 +1,5 @@
 import { Component, OnInit } from '@angular/core'
-import { BootstrapService, ColorService } from '@lib/common'
-import { SearchFacade } from '@lib/search'
-import { map, pluck, take, tap } from 'rxjs/operators'
+import { ColorService } from '@lib/common'
 
 @Component({
   selector: 'app-root',
@@ -11,24 +9,9 @@ import { map, pluck, take, tap } from 'rxjs/operators'
 export class AppComponent implements OnInit {
   title = 'search'
 
-  constructor(
-    private bootstrap: BootstrapService,
-    private searchFacade: SearchFacade
-  ) {
+  constructor() {
     ColorService.applyCssVariables('#e73f51', '#c2e9dc', '#212029', '#fdfbff')
   }
 
-  ngOnInit(): void {
-    this.bootstrap
-      .uiConfReady('srv')
-      .pipe(
-        take(1),
-        map((config) => config.mods.search.facetConfig),
-        tap((aggregationsConfig) => {
-          this.searchFacade.setConfigAggregations(aggregationsConfig)
-          this.searchFacade.requestMoreResults()
-        })
-      )
-      .subscribe()
-  }
+  ngOnInit(): void {}
 }
