+++ apps/search/src/app/main-search/main-search.component.html
@@ -1,0 +1,18 @@
+<div class="mb-3 mx-4 rounded flex flex-row">
+  <div class="w-2/4 mr-3">
+    <search-fuzzy-search></search-fuzzy-search>
+  </div>
+  <div class="w-1/4 mr-3">
+    <search-sort-by></search-sort-by>
+  </div>
+  <div class="flex-grow">
+    <search-results-layout></search-results-layout>
+  </div>
+</div>
+<div class="flex flex-row">
+  <search-facets-container class="w-1/4"></search-facets-container>
+  <div class="w-3/4">
+    <search-results-hits></search-results-hits>
+    <search-results-list-container></search-results-list-container>
+  </div>
+</div>
